Refcator FetchLoader as a HOC and compose containers in index.js
