+++ src/containers/HcaLandingPageContainer.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import URI from 'urijs'
 
-import CalloutAlert from './CalloutAlert'
+import withFetchLoader from './FetchLoader'
+import cardPropTypes from '../cards/cardPropTypes'
 import ExtendableCard from '../cards/ExtendableCard'
 
 const CardContainer = styled.div`
@@ -13,83 +13,20 @@ const CardContainer = styled.div`
   }
 `
 
-class HcaCardContainer extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      data: null,
-      isLoading: true,
-      hasError: null
+const HcaCardContainer = ({cards}) =>
+  <div className={`row`}>
+    {
+      Array.isArray(cards) &&
+      cards.map((card, index) =>
+        <CardContainer className={`small-3 small-centered`} key={index}>
+          <ExtendableCard {...card} />
+        </CardContainer>
+      )
     }
-  }
-
-  async componentDidMount() {
-    this.setState({ isLoading: true })
-
-    const url = URI(this.props.resource, this.props.host).toString()
-
-    try {
-      const response = await fetch(url)
-
-      if (!response.ok) {
-        throw new Error(`${url} => ${response.status}`)
-      }
-
-      this.setState({
-        data: await response.json(),
-        isLoading: false,
-        hasError: null
-      })
-    } catch(e) {
-      this.setState({
-        data: null,
-        isLoading: false,
-        hasError: {
-          description: `There was a problem communicating with the server. Please try again later.`,
-          name: e.name,
-          message: e.message
-        }
-      })
-    }
-  }
-
-  componentDidCatch(error, info) {
-    this.setState({
-      hasError: {
-        description: `There was a problem rendering this component.`,
-        name: error.name,
-        message: `${error.message} – ${info}`
-      }
-    })
-  }
-
-  render() {
-    const { data, isLoading, hasError } = this.state
-
-    return (
-      hasError ?
-        <CalloutAlert error={hasError} /> :
-      isLoading ?
-        <p className={`row column`} id={`loading-message`}> Loading, please wait...</p> :
-      // Promise fulfilled
-        <div className={`row`}>
-          {
-            Array.isArray(data) &&
-            data.map((card, index) =>
-              <CardContainer className={`small-3 small-centered`} key={index}>
-                <ExtendableCard {...card} />
-              </CardContainer>
-            )
-          }
-        </div>
-    )
-  }
-}
+  </div>
 
 HcaCardContainer.propTypes = {
-  host: PropTypes.string.isRequired,
-  resource: PropTypes.string.isRequired
+  cards: PropTypes.arrayOf(PropTypes.shape(cardPropTypes)).isRequired
 }
 
 export default HcaCardContainer
