+++ src/index.js
@@ -1,4 +1,11 @@
+import withFetchLoader from './containers/FetchLoader'
 import SceaHomepageSpeciesContainer from './containers/SceaHomepageSpeciesContainer'
 import HcaLandingPageContainer from './containers/HcaLandingPageContainer'
 
-export { SceaHomepageSpeciesContainer, HcaLandingPageContainer }
+const _SceaHomepageSpeciesContainer = withFetchLoader(SceaHomepageSpeciesContainer)
+const _HcaLandingPageContainer = withFetchLoader(HcaLandingPageContainer)
+
+export {
+  _SceaHomepageSpeciesContainer as SceaHomepageSpeciesContainer,
+  _HcaLandingPageContainer as HcaLandingPageContainer
+}
